♻️ Adding donation and origin
