+++ components/Bowswap/InputTokenDisabled.js
@@ -3,23 +3,23 @@ import	{ethers}			from	'ethers';
 
 function	InputTokenDisabled({
 	value, toCounterValue,
-	slippage, donation,
-	balanceOf, decimals,
+	slippage, balanceOf, decimals,
 	isFetchingExpectedReceiveAmount
 }) {
-	const	expectedOutWithSettings = Number(value) !== 0 ? (Number(value) - ((Number(value) * (slippage + donation) / 100))) : 0;
+	const	expectedOutWithSettings = Number(value) !== 0 ? (Number(value) - ((Number(value) * slippage / 100))) : 0;
 	const	tokenBalance = Number(ethers.utils.formatUnits(balanceOf, decimals));
+
 	return (
 		<div className={'relative w-full text-left bg-white border border-ygray-100 rounded-lg cursor-default focus:outline-none flex flex-col justify-between text-ygray-800 h-24 py-2 px-2 space-y-1'}>
 			<div className={'h-4'}>
 				<div className={'flex flex-row items-center justify-end w-full'}>
 					<label
 						className={'font-normal text-ybase text-ygray-500 hidden md:flex flex-row items-center cursor-pointer'}>
-						{`Balance: ${tokenBalance < 1 ? '< 1' : tokenBalance}`}
+						{`Balance: ${tokenBalance !== 0 && tokenBalance < 1 ? '< 1' : tokenBalance}`}
 					</label>
 					<label
 						className={'font-normal text-ybase text-ygray-500 flex flex-row items-center md:hidden cursor-pointer'}>
-						{`Balance: ${tokenBalance < 1 ? '< 1' : tokenBalance.toFixed(8)}`}
+						{`Balance: ${tokenBalance !== 0 && tokenBalance < 1 ? '< 1' : tokenBalance.toFixed(8)}`}
 					</label>
 				</div>
 			</div>
