+++ components/Bowswap/index.js
@@ -4,7 +4,6 @@ import	useWeb3											from	'contexts/useWeb3';
 import	useAccount										from	'contexts/useAccount';
 import	useLocalStorage									from	'hook/useLocalStorage';
 import	useDebounce										from	'hook/useDebounce';
-import	{approveToken, metapoolSwapTokens, swapTokens, metapoolSwapTokensWithSignature, swapTokensWithSignature, signTransaction}	from	'utils/actions';
 import	InputToken										from	'components/Bowswap/InputToken';
 import	InputTokenDisabled								from	'components/Bowswap/InputTokenDisabled';
 import	ModalVaultList									from	'components/Bowswap/ModalVaultList';
@@ -18,6 +17,7 @@ import	BOWSWAP_CRV_USD_VAULTS							from	'utils/BOWSWAP_CRV_USD_VAULTS';
 import	BOWSWAP_CRV_V2_VAULTS							from	'utils/BOWSWAP_CRV_V2_VAULTS';
 import	V2_PATHS										from	'utils/currentPaths';
 import	{bigNumber, toAddress}							from	'utils';
+import SectionButtons from './SectionButtons';
 
 function	SectionFromVault({
 	vaults, fromVault, set_fromVault,
@@ -94,8 +94,7 @@ function	SectionFromVault({
 function	SectionToVault({
 	vaults, toVault, set_toVault,
 	expectedReceiveAmount, toCounterValue, balanceOf,
-	slippage, donation,
-	isFetchingExpectedReceiveAmount, disabled,
+	slippage, isFetchingExpectedReceiveAmount, disabled,
 	yearnVaultData
 }) {
 	return (
@@ -117,7 +116,6 @@ function	SectionToVault({
 						value={expectedReceiveAmount}
 						toCounterValue={toCounterValue}
 						slippage={slippage}
-						donation={donation}
 						balanceOf={balanceOf}
 						decimals={toVault.decimals}
 						isFetchingExpectedReceiveAmount={isFetchingExpectedReceiveAmount} />
@@ -127,335 +125,9 @@ function	SectionToVault({
 	);
 }
 
-function	ButtonSwap({
-	fromVault, toVault,
-	fromAmount, expectedReceiveAmount,
-	slippage, donation,
-	signature, shouldIncreaseGasLimit,
-	approved, disabled, onCallback
-}) {
-	const	{provider} = useWeb3();
-	const	[transactionProcessing, set_transactionProcessing] = useState(false);
-
-	const	[DEBUG_TX, set_DEBUG_TX] = useState(-1);
-
-	useEffect(() => {
-		window.swap = () => set_DEBUG_TX(n => n + 1);
-	}, [typeof(window) !== 'undefined']);
-
-	useEffect(() => {
-		if (DEBUG_TX >= 0)
-			performSwap(true);
-	}, [DEBUG_TX]);
-
-	function	performV2Swap() {
-		try {
-			swapTokens({
-				provider: provider,
-				contractAddress: process.env.METAPOOL_SWAPPER_ADDRESS,
-				from: fromVault.address,
-				to: toVault.address,
-				amount: ethers.utils.parseUnits(fromAmount, fromVault.decimals),
-				minAmountOut: ethers.utils.parseUnits((expectedReceiveAmount - (expectedReceiveAmount * (slippage + donation) / 100)).toString(), fromVault.decimals),
-				instructions: V2_PATHS.find(path => path[0] === fromVault.address && path[1] === toVault.address)?.[2],
-				donation,
-				shouldIncreaseGasLimit
-			}, ({error}) => {
-				if (error) {
-					let message = undefined;
-					if (error?.data?.message?.includes('revert out too low')) {
-						message = 'SLIPPAGE TOO HIGH. TO PROCEED, PLEASE INCREASE THE SLIPPAGE TOLERANCE';
-					}
-					set_transactionProcessing(false);
-					return onCallback('error', message);
-				}
-				set_transactionProcessing(false);
-				onCallback('success');
-			});
-		} catch (error) {
-			let message = undefined;
-			if (error?.data?.message?.includes('revert out too low')) {
-				message = 'SLIPPAGE TOO HIGH. TO PROCEED, PLEASE INCREASE THE SLIPPAGE TOLERANCE';
-			}
-			set_transactionProcessing(false);
-			return onCallback('error', message);
-		}
-	}
-	function	performV1Swap() {
-		const	v2PathExists = V2_PATHS.find(path => path[0] === fromVault.address && path[1] === toVault.address);
-
-		try {
-			metapoolSwapTokens({
-				provider: provider,
-				contractAddress: process.env.METAPOOL_SWAPPER_ADDRESS,
-				from: fromVault.address,
-				to: toVault.address,
-				amount: ethers.utils.parseUnits(fromAmount, fromVault.decimals),
-				minAmountOut: ethers.utils.parseUnits((expectedReceiveAmount - (expectedReceiveAmount * (slippage + donation) / 100)).toString(), fromVault.decimals),
-				donation,
-				shouldIncreaseGasLimit
-			}, ({error}) => {
-				if (error) {
-					if (error?.message?.includes('User denied transaction signature')) {
-						set_transactionProcessing(false);
-						return onCallback('error', 'User denied transaction signature');
-					} else if (v2PathExists) {
-						console.log('FALLBACK_WITH_V2');
-						return performV2Swap();
-					} else {
-						let message = undefined;
-						if (error?.data?.message?.includes('revert out too low')) {
-							message = 'SLIPPAGE TOO HIGH. TO PROCEED, PLEASE INCREASE THE SLIPPAGE TOLERANCE';
-						}
-						set_transactionProcessing(false);
-						return onCallback('error', message);
-					}
-				}
-				set_transactionProcessing(false);
-				onCallback('success');
-			});
-		} catch (error) {
-			if (error?.message?.includes('User denied transaction signature')) {
-				set_transactionProcessing(false);
-				return onCallback('error', 'User denied transaction signature');
-			} else if (v2PathExists) {
-				console.log('FALLBACK_WITH_V2');
-				return performV2Swap();
-			} else {
-				let message = undefined;
-				if (error?.data?.message?.includes('revert out too low')) {
-					message = 'SLIPPAGE TOO HIGH. TO PROCEED, PLEASE INCREASE THE SLIPPAGE TOLERANCE';
-				}
-				set_transactionProcessing(false);
-				return onCallback('error', message);
-			}
-		}
-	}
-
-	function	performV2SwapWithSignature() {
-		swapTokensWithSignature({
-			provider: provider,
-			contractAddress: process.env.SIGNATURE_METAPOOL_SWAPPER_ADDRESS,
-			from: fromVault.address,
-			to: toVault.address,
-			amount: ethers.utils.parseUnits(fromAmount, fromVault.decimals),
-			minAmountOut: ethers.utils.parseUnits((expectedReceiveAmount - (expectedReceiveAmount * (slippage + donation) / 100)).toString(), fromVault.decimals),
-			instructions: V2_PATHS.find(path => path[0] === fromVault.address && path[1] === toVault.address)?.[2],
-			signature,
-			donation,
-			shouldIncreaseGasLimit
-		}, ({error}) => {
-			if (error) {
-				if (error?.message?.includes('User denied transaction signature')) {
-					set_transactionProcessing(false);
-					return onCallback('error', 'User denied transaction signature');
-				} else {
-					let message = undefined;
-					if (error?.data?.message?.includes('revert out too low')) {
-						message = 'SLIPPAGE TOO HIGH. TO PROCEED, PLEASE INCREASE THE SLIPPAGE TOLERANCE';
-					} else if (error?.data?.message?.includes('execution reverted')) {
-						message = 'INVALID SIGNATURE, FALLBACK TO DEFAULT FLOW';
-					} else {
-						message = 'INVALID SIGNATURE, FALLBACK TO DEFAULT FLOW';
-					}
-					set_transactionProcessing(false);
-					return onCallback('error', message);
-				}
-			}
-			set_transactionProcessing(false);
-			onCallback('success');
-		});
-	}
-	function	performV1SwapWithSignature() {
-		const	v2PathExists = V2_PATHS.find(path => path[0] === fromVault.address && path[1] === toVault.address);
-
-		try {
-			metapoolSwapTokensWithSignature({
-				provider: provider,
-				contractAddress: process.env.SIGNATURE_METAPOOL_SWAPPER_ADDRESS,
-				from: fromVault.address,
-				to: toVault.address,
-				amount: ethers.utils.parseUnits(fromAmount, fromVault.decimals),
-				minAmountOut: ethers.utils.parseUnits((expectedReceiveAmount - (expectedReceiveAmount * (slippage + donation) / 100)).toString(), fromVault.decimals),
-				signature,
-				donation,
-				shouldIncreaseGasLimit
-			}, ({error}) => {
-				if (error) {
-					if (error?.message?.includes('User denied transaction signature')) {
-						set_transactionProcessing(false);
-						return onCallback('error', 'User denied transaction signature');
-					} else if (v2PathExists) {
-						console.log('FALLBACK_WITH_V2');
-						return performV2SwapWithSignature();
-					} else {
-						let message = undefined;
-						if (error?.data?.message?.includes('revert out too low')) {
-							message = 'SLIPPAGE TOO HIGH. TO PROCEED, PLEASE INCREASE THE SLIPPAGE TOLERANCE';
-						} else if (error?.message?.includes('execution reverted')) {
-							message = 'INVALID SIGNATURE, FALLBACK TO DEFAULT FLOW';
-						} else {
-							message = 'INVALID SIGNATURE, FALLBACK TO DEFAULT FLOW';
-						}
-						set_transactionProcessing(false);
-						return onCallback('error', message);
-					}
-				}
-				set_transactionProcessing(false);
-				onCallback('success');
-			});
-		} catch (error) {
-			if (error?.message?.includes('User denied transaction signature')) {
-				set_transactionProcessing(false);
-				return onCallback('error', 'User denied transaction signature');
-			} else if (v2PathExists) {
-				console.log('FALLBACK_WITH_V2');
-				return performV2Swap();
-			} else {
-				let message = undefined;
-				if (error?.message?.includes('revert out too low')) {
-					message = 'SLIPPAGE TOO HIGH. TO PROCEED, PLEASE INCREASE THE SLIPPAGE TOLERANCE';
-				}
-				set_transactionProcessing(false);
-				return onCallback('error', message);
-			}
-		}
-	}
-
-	function	performSwap(forced = false) {
-		if (!forced && (disabled || transactionProcessing || !approved)) {
-			return;
-		}
-		set_transactionProcessing(true);
-		onCallback('pending');
-		if (toVault.scope === 'v2') {
-			if (signature) {
-				return performV2SwapWithSignature();
-			}
-			performV2Swap();
-		} else {
-			if (signature) {
-				return performV1SwapWithSignature();
-			}
-			performV1Swap();
-		}
-	}
-
-	return (
-		<button
-			onClick={() => performSwap(false)}
-			className={`w-full h-11 flex items-center justify-center space-x-2 px-6 py-3 text-ybase font-medium rounded-lg focus:outline-none overflow-hidden transition-colors border ${
-				disabled || transactionProcessing || !approved ? 'text-ygray-400 bg-white border-ygray-400 cursor-not-allowed' :
-					'bg-yblue hover:bg-yblue-hover border-yblue hover:border-yblue-hover text-white cursor-pointer'
-			}`}>
-			<span>{'Swap'}</span>
-		</button>
-	);
-}
-
-function	ButtonApprove({fromVault, fromAmount, approved, disabled, set_signature, canSign, onCallback}) {
-	const	{provider} = useWeb3();
-	const	[transactionProcessing, set_transactionProcessing] = useState(false);
-	const	[DEBUG_TX, set_DEBUG_TX] = useState(-1);
-
-	useEffect(() => {
-		window.approve = (nonce) => set_DEBUG_TX(nonce);
-	}, [typeof(window) !== 'undefined']);
-
-	useEffect(() => {
-		if (DEBUG_TX >= 0)
-			performApprove(true, DEBUG_TX);
-	}, [DEBUG_TX]);
-
-	function	approveTx() {
-		try {
-			approveToken({
-				provider: provider,
-				contractAddress: fromVault.address,
-				amount: ethers.utils.parseUnits(fromAmount, fromVault.decimals),
-				from: process.env.METAPOOL_SWAPPER_ADDRESS
-			}, ({error}) => {
-				if (error) {
-					set_transactionProcessing(false);
-					return onCallback('error');
-				}
-				set_transactionProcessing(false);
-				onCallback('success');
-			});	
-		} catch (error) {
-			set_transactionProcessing(false);
-			onCallback('error');
-		}
-	}
-
-	function	performApprove(forced = false, nonceOverwrite = undefined) {
-		if (!forced && (disabled || transactionProcessing || (!fromAmount || Number(fromAmount) === 0))) {
-			return;
-		}
-		set_transactionProcessing(true);
-		onCallback('pending');
-		if (!canSign) {
-			return approveTx();
-		}
-		try {
-			signTransaction({
-				provider: provider,
-				vaultAddress: fromVault.address,
-				contractAddress: process.env.SIGNATURE_METAPOOL_SWAPPER_ADDRESS,
-				amount: ethers.utils.parseUnits(fromAmount, fromVault.decimals),
-				nonceOverwrite 
-			}, ({error, data}) => {
-				if (error) {
-					if (error?.message?.includes('User denied message signature')) {
-						set_transactionProcessing(false);
-						return onCallback('error', 'User denied signature');
-					}
-					console.log('FALLBACK_WITH_APPROVE');
-					return approveTx();
-				}
-				set_signature(data);
-				set_transactionProcessing(false);
-				onCallback('success');
-			});
-		} catch (error) {
-			if (error?.message?.includes('User denied message signature')) {
-				set_transactionProcessing(false);
-				return onCallback('error', 'User denied signature');
-			}
-			console.log('FALLBACK_WITH_APPROVE');
-			return approveTx();
-		}
-	}
-
-	if (approved) {
-		return (
-			<button
-				className={'w-full h-11 flex items-center justify-center space-x-2 px-6 py-3 text-ybase font-medium rounded-lg focus:outline-none overflow-hidden transition-colors border bg-white border-yblue text-yblue cursor-not-allowed'}>
-				<div className={'flex flex-row items-center justify-center'}>
-					<span>{'Approved'}</span>
-					<svg className={'ml-2'} width={'16'} height={'16'} viewBox={'0 0 16 16'} fill={'none'} xmlns={'http://www.w3.org/2000/svg'}>
-						<path fillRule={'evenodd'} clipRule={'evenodd'} d={'M13.7602 3.48787C14.043 3.72358 14.0813 4.14396 13.8456 4.42681L7.17889 12.4268C6.96078 12.6885 6.58042 12.7437 6.29694 12.5547L2.29694 9.88805C1.99058 9.68382 1.9078 9.2699 2.11204 8.96355C2.31627 8.6572 2.73019 8.57442 3.03654 8.77865L6.53809 11.113L12.8213 3.57323C13.057 3.29038 13.4773 3.25216 13.7602 3.48787Z'} fill={'#00A3FF'}/>
-					</svg>
-				</div>
-			</button>
-		);
-	}
-
-	return (
-		<button
-			onClick={performApprove}
-			className={`w-full h-11 flex items-center justify-center space-x-2 px-6 py-3 text-ybase font-medium rounded-lg focus:outline-none overflow-hidden transition-colors border ${
-				disabled || transactionProcessing || (!fromAmount || Number(fromAmount) === 0) ? 'text-ygray-400 bg-white border-ygray-400 cursor-not-allowed' :
-					'bg-yblue border-yblue hover:bg-yblue-hover hover:border-yblue-hover text-white cursor-pointer'
-			}`}>
-			<span>{'Approve'}</span>
-		</button>
-	);
-}
-
 function	Bowswap({prices}) {
 	const	{provider} = useWeb3();
-	const	{balancesOf, updateBalanceOf, allowances, yearnVaultData} = useAccount();
+	const	{balancesOf, yearnVaultData} = useAccount();
 	const	[, set_nonce] = useState(0);
 	const	[fromVault, set_fromVault] = useLocalStorage('fromVault', BOWSWAP_CRV_USD_VAULTS[0]);
 	const	[fromCounterValue, set_fromCounterValue] = useLocalStorage('fromCounterValue', 0);
@@ -473,7 +145,6 @@ function	Bowswap({prices}) {
 	const	[txApproveStatus, set_txApproveStatus] = useState({none: true, pending: false, success: false, error: false});
 	const	[txSwapStatus, set_txSwapStatus] = useState({none: true, pending: false, success: false, error: false});
 	const	[signature, set_signature] = useState(null);
-	const	[canSign, set_canSign] = useState(true);
 
 	function	resetStates() {
 		set_signature(null);
@@ -495,7 +166,7 @@ function	Bowswap({prices}) {
 	}
 
 	const	fetchEstimateOut = useCallback(async (from, to, amount) => {
-		const	fromToken = new ethers.Contract(process.env.METAPOOL_SWAPPER_ADDRESS, [
+		const	fromToken = new ethers.Contract(process.env.BOWSWAP_SWAPPER_ADDR, [
 			'function metapool_estimate_out(address from, address to, uint256 amount, uint256 donation) public view returns (uint256)',
 			'function estimate_out(address from, address to, uint256 amount, tuple(bool deposit, address pool, uint128 n)[] instructions, uint256 donation) public view returns (uint256)'
 		], provider);
@@ -507,18 +178,23 @@ function	Bowswap({prices}) {
 					to,
 					amount,
 					V2_PATHS.find(path => path[0] === fromVault.address && path[1] === toVault.address)?.[2],
-					donation
+					10_000 - donation * 100
 				);
 				set_expectedReceiveAmount(ethers.utils.formatUnits(estimate_out, toVault.decimals));
 				set_isFetchingExpectedReceiveAmount(false);
 			}
 		} else {
-			const	metapool_estimate_out = await fromToken.metapool_estimate_out(from, to, amount, donation);
+			const	metapool_estimate_out = await fromToken.metapool_estimate_out(
+				from,
+				to,
+				amount,
+				10_000 - donation * 100
+			);
 			set_expectedReceiveAmount(ethers.utils.formatUnits(metapool_estimate_out, toVault.decimals));
 			set_isFetchingExpectedReceiveAmount(false);
 		}
 	// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [fromVault.address, provider, toVault?.address, toVault?.decimals, toVault?.scope, toVault?.type]);
+	}, [fromVault.address, provider, toVault?.address, toVault?.decimals, toVault?.scope, toVault?.type, donation]);
 
 	/**************************************************************************
 	**	This function will be used to compute the counter value of the want
@@ -657,7 +333,6 @@ function	Bowswap({prices}) {
 	// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [toVault?.address, provider]);
 
-
 	/**************************************************************************
 	**	This is used to get the estimateOut based on the current inputed
 	**	amount. This is triggered with a delay to avoid too many requests
@@ -674,7 +349,7 @@ function	Bowswap({prices}) {
 			set_expectedReceiveAmount('');
 		}
 	// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [debouncedFetchExpectedAmount, fromAmount, fromVault.address, toVault?.address, fromVault.decimals]);
+	}, [debouncedFetchExpectedAmount, fromAmount, donation, fromVault.address, toVault?.address, fromVault.decimals]);
 
 
 	function	renderMiddlePart() {
@@ -725,7 +400,6 @@ function	Bowswap({prices}) {
 		);
 	}
 
-	const	isAllowed = fromAmount !== '' && Number(fromAmount) > 0 && ethers.BigNumber.from(allowances?.[fromVault.address] || 0)?.gte(ethers.utils.parseUnits(fromAmount || '0', fromVault.decimals || 18));
 	return (
 		<div className={'w-full max-w-2xl'}>
 			<div className={'bg-white rounded-xl shadow-base p-4 w-full relative space-y-0 md:space-y-4'}>
@@ -756,65 +430,24 @@ function	Bowswap({prices}) {
 					expectedReceiveAmount={expectedReceiveAmount}
 					toCounterValue={toCounterValue}
 					slippage={slippage}
-					donation={donation}
 					balanceOf={balancesOf[toVault.address]?.toString() || '0'}
 					isFetchingExpectedReceiveAmount={isFetchingExpectedReceiveAmount}
 					yearnVaultData={yearnVaultData} />
 
-				<div className={'flex flex-row justify-center pt-8 w-full space-x-4'}>
-					<ButtonApprove
-						disabled={(Number(fromAmount) > Number(ethers.utils.formatUnits(balancesOf[fromVault.address]?.toString() || '0', fromVault.decimals)) || txApproveStatus.success || isAllowed )}
-						approved={txApproveStatus.success || isAllowed}
-						fromVault={fromVault}
-						fromAmount={fromAmount}
-						set_signature={set_signature}
-						canSign={canSign}
-						onCallback={(type, message) => {
-							set_txApproveStatus({none: false, pending: type === 'pending', error: type === 'error', success: type === 'success', message});
-							if (type === 'error') {
-								setTimeout(() => set_txApproveStatus((s) => s.error ? {none: true, pending: false, error: false, success: false, message} : s), 2500);
-							}
-							if (type === 'success') {
-								updateBalanceOf([fromVault.address]);
-								setTimeout(() => set_txApproveStatus({none: false, pending: false, error: false, success: true, hide: true, message: null}), 2500);
-							}
-						}} />
-					<ButtonSwap
-						disabled={Number(fromAmount) > Number(ethers.utils.formatUnits(balancesOf[fromVault.address]?.toString() || '0', fromVault.decimals))}
-						approved={txApproveStatus.success || isAllowed || signature}
-						fromVault={fromVault}
-						toVault={toVault}
-						fromAmount={fromAmount}
-						expectedReceiveAmount={expectedReceiveAmount}
-						slippage={slippage}
-						donation={donation}
-						signature={signature}
-						shouldIncreaseGasLimit={Number(balanceOfFromVault) < Number(fromAmount)}
-						onCallback={(type, message) => {
-							set_txApproveStatus({none: false, pending: false, error: false, success: true, hide: true, message: null});
-
-							set_txSwapStatus({none: false, pending: type === 'pending', error: type === 'error', success: type === 'success', message});
-							if (type === 'error') {
-								if (message === 'INVALID SIGNATURE, FALLBACK TO DEFAULT FLOW') {
-									setTimeout(() => {
-										set_canSign(false);
-										set_signature(null);
-										set_txApproveStatus({none: true, pending: false, error: false, success: false, message: undefined});
-										set_txSwapStatus({none: true, pending: false, error: false, success: false});
-									}, 2500);
-
-								} else {
-									setTimeout(() => set_txSwapStatus((s) => s.error ? {none: true, pending: false, error: false, success: false, message} : s), 2500);
-								}
-							}
-							if (type === 'success') {
-								updateBalanceOf([fromVault.address, toVault.address]);
-								setTimeout(() => set_txSwapStatus({none: true, pending: false, error: false, success: false}), 2500);
-								resetStates();
-							}
-						}}
-					/>
-				</div>
+				<SectionButtons
+					fromAmount={fromAmount}
+					fromVault={fromVault}
+					balanceOfFromVault={balanceOfFromVault}
+					expectedReceiveAmount={expectedReceiveAmount}
+					toVault={toVault}
+					slippage={slippage}
+					donation={donation}
+					txApproveStatus={txApproveStatus}
+					set_txApproveStatus={set_txApproveStatus}
+					signature={signature}
+					set_signature={set_signature}
+					set_txSwapStatus={set_txSwapStatus}
+					resetStates={resetStates} />
 			</div>
 		</div>
 	);
