+++ next.config.js
@@ -6,9 +6,8 @@ module.exports = ({
 	],
 	env: {
 		ALCHEMY_KEY: process.env.ALCHEMY_KEY,
-		SIGNATURE_METAPOOL_SWAPPER_ADDRESS: '0xF12eeAB1C759dD7D8C012CcA6d8715EEd80e51b6',
-		METAPOOL_SWAPPER_ADDRESS: '0xF12eeAB1C759dD7D8C012CcA6d8715EEd80e51b6',
-		VYEMPIRE_SWAPPER: '0xEB8D98f9E42a15b0Eb35315F737bdfDa1a8D2Eaa',
+		BOWSWAP_SWAPPER_ADDR: '0x00000000ff0027b3f3d2b0412d4a4ced4e3366da',
+		VYEMPIRE_SWAPPER_ADDR: '0xEB8D98f9E42a15b0Eb35315F737bdfDa1a8D2Eaa',
 		SECRET: process.env.SECRET,
 	},
 	optimization: {
