+++ readme.md
@@ -6,7 +6,6 @@
 2. Create a `.env` file in the root directory with
 ```
 ALCHEMY_KEY=YOUR_ALCHEMY_KEY
-METAPOOL_SWAPPER_ADDRESS=ADDRESS_OF_THE_METAPOOL_SWAPPER
 ```  
 3. Start the dev project with `yarn run dev`
 4. [Go check the swap](http://localhost:3000)
